pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend for RAG chatbot with OpenAI Agents SDK and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "openai-agents>=0.1.0",
    "openai-agents[litellm]>=0.1.0",
    "cohere>=5.5.3",
    "qdrant-client>=1.8.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "google-generativeai>=0.4.0",
    "beautifulsoup4>=4.12.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]